# docs/swagger.yaml
definitions:
...
  fiber.Map:
    additionalProperties: true
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.CareerStage:
    enum:
    - entrylevel
    - senior
    type: string
    x-enum-varnames:
    - CareerStageEntryLevel
    - CareerStageSenior
  models.Industry:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      industry:
        type: string
      organization:
        items:
          $ref: '#/definitions/models.Organization'
        type: array
      updatedAt:
        type: string
    type: object
  models.Media:
    enum:
    - website
    - facebook
    - instagram
    - tiktok
    - youtube
    - linkedin
    - line
    type: string
    x-enum-varnames:
    - MediaWebsite
    - MediaFacebook
    - MediaIG
    - MediaTikTok
    - MediaYoutube
    - MediaLinkedin
    - MediaLine
  models.OrgOpenJob:
    properties:
      benefits:
        example: Health insurance
        type: string
      career_stage:
        allOf:
        - $ref: '#/definitions/models.CareerStage'
        example: junior
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        example: This is a description
        type: string
      hours_per_day:
        example: 8 hours
        type: string
      id:
        example: 1
        type: integer
      organization:
        example: builds CMU
        type: string
      organization_id:
        example: 1
        type: integer
      period:
        example: 1 year
        type: string
      prerequisite:
        description: Required qualifications or skills
        example:
        - Great at problem solving
        - Reliable
        items:
          type: string
        type: array
      qualifications:
        example: Bachelor's degree in Computer Science
        type: string
      quantity:
        example: 1
        type: integer
      salary:
        example: 30000
        type: number
      scope:
        example: Software Development
        type: string
      title:
        example: Software Engineer
        type: string
      updatedAt:
        type: string
      work_type:
        allOf:
        - $ref: '#/definitions/models.WorkType'
        example: fulltime
      workplace:
        allOf:
        - $ref: '#/definitions/models.Workplace'
        example: remote
    type: object
  models.Organization:
    properties:
      created_at:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      expertise:
        description: Organization's area of expertise
        type: string
      goal:
        description: Detailed description of the organization's goal
        items:
          type: string
        type: array
      id:
        type: integer
      industry:
        items:
          $ref: '#/definitions/models.Industry'
        type: array
      latitude:
        description: Geographic latitude (stored as string for precision)
        type: string
      location:
        description: General location
        type: string
      longitude:
        description: Geographic longitude (stored as string for precision)
        type: string
      org_email:
        description: Email address (unique constraint)
        type: string
      org_name:
        type: string
      org_phone:
        type: string
      orgOpenJobs:
        items:
          $ref: '#/definitions/models.OrgOpenJob'
        type: array
      organizationContacts:
        items:
          $ref: '#/definitions/models.OrganizationContact'
        type: array
      pic_url:
        description: URL to organization's logo
        type: string
      postal_code:
        description: Postal code, allowing for flexibility in format
        type: string
      province:
        description: Province name
        type: string
      subdistrict:
        description: Subdistrict name
        type: string
      updated_at:
        type: string
      updatedAt:
        type: string
    type: object
  models.OrganizationContact:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      media:
        $ref: '#/definitions/models.Media'
      media_link:
        type: string
      organization_id:
        type: integer
      updatedAt:
        type: string
    type: object
...
  models.WorkType:
    enum:
    - fulltime
    - parttime
    - internship
    - volunteer
    type: string
    x-enum-varnames:
    - WorkTypeFullTime
    - WorkTypePartTime
    - WorkTypeInternship
    - WorkTypeVolunteer
  models.Workplace:
    enum:
    - onsite
    - remote
    - hybrid
    type: string
    x-enum-varnames:
    - WorkplaceOnsite
    - WorkplaceRemote
    - WorkplaceHybrid
...
info:
  contact: {}
paths:
...
  /orgs:
    post:
      consumes:
      - application/json
      description: Create a new organization BUT still not create the Contact and
        OpenJob
      parameters:
      - description: Organization
        in: body
        name: org
        required: true
        schema:
          $ref: '#/definitions/models.Organization'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Organization'
        "400":
          description: Bad Request - json body is required or invalid / organization
            name is required
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Create a new organization
      tags:
      - Organization
  /orgs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request - organization id is required / invalid organization
            id
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Delete an organization by ID
      tags:
      - Organization
    get:
      consumes:
      - application/json
      description: Get an organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Organization'
        "400":
          description: Bad Request - organization id is required
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Not Found - organization not found
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Get an organization by ID
      tags:
      - Organization
    put:
      consumes:
      - application/json
      description: Update an organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      - description: Organization
        in: body
        name: org
        required: true
        schema:
          $ref: '#/definitions/models.Organization'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Organization'
        "400":
          description: Bad Request - organization id is required / invalid organization
            ID
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Update an organization by ID
      tags:
      - Organization
...
  /orgs/{orgID}/jobs/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an organization job by ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request - organization id & job id is required
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Delete an organization job by ID
      tags:
      - Organization Job
  /orgs/{orgID}/jobs/get/{id}:
    get:
      consumes:
      - application/json
      description: Get an organization open job by ID
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: integer
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrgOpenJob'
        "400":
          description: Bad Request - organization id & job id is required
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Not Found - jobs not found
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Get an organization open job by ID
      tags:
      - Organization Job
  /orgs/{orgID}/jobs/list:
    get:
      consumes:
      - application/json
      description: Get all organization open jobs
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OrgOpenJob'
            type: array
        "400":
          description: Bad Request - organization id is required or invalid
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: List all jobs of its organization
      tags:
      - Organization Job
  /orgs/{orgID}/jobs/open:
    post:
      consumes:
      - application/json
      description: Create a new organization open job
      parameters:
      - description: Organization Open Job
        in: body
        name: org
        required: true
        schema:
          $ref: '#/definitions/models.OrgOpenJob'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OrgOpenJob'
        "400":
          description: Bad Request - json body is required or invalid / job title
            is required
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Create a new organization open job
      tags:
      - Organization Job
  /orgs/{orgID}/jobs/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an organization open job by ID
      parameters:
      - description: Organization Open Job
        in: body
        name: org
        required: true
        schema:
          $ref: '#/definitions/models.OrgOpenJob'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrgOpenJob'
        "400":
          description: Bad Request - organization id & job id is required
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Update an organization open job by ID
      tags:
      - Organization Job
  /orgs/jobs/list/all:
    get:
      consumes:
      - application/json
      description: Get all organization jobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OrgOpenJob'
            type: array
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: List all organization jobs
      tags:
      - Organization Job
  /orgs/list:
    get:
      consumes:
      - application/json
      description: Get all organizations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Organization'
            type: array
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: List all organizations
      tags:
      - Organization
  /orgs/paginate:
    get:
      consumes:
      - application/json
      description: Get a page of organizations
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Organization'
            type: array
        "400":
          description: Bad Request - invalid page
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Get a page of organizations
      tags:
      - Organization
...
swagger: "2.0"